+++ plugins/pagefind.ts
@@ -90,7 +90,7 @@ const defaults: Options = {
   binary: {
     path: "./_bin/pagefind",
     extended: false,
-    version: "v0.10.6",
+    version: "v0.10.7",
   },
   ui: {
     containerId: "search",