+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project try to adheres to [Semantic Versioning](https://semver.org/),
 but not always is possible (due the use of unstable features from Deno).
 Any BREAKING CHANGE between minor versions will be documented here in upper case.
 
-## [Unreleased]
+## [1.15.2] - 2023-02-02
 ### Added
 - Property `inheritData` for components [#364].
 - Support for import maps in the deno.json file.
@@ -2077,7 +2077,7 @@ The first version.
 [#359]: https://github.com/lumeland/lume/issues/359
 [#364]: https://github.com/lumeland/lume/issues/364
 
-[Unreleased]: https://github.com/lumeland/lume/compare/v1.15.1...HEAD
+[1.15.2]: https://github.com/lumeland/lume/compare/v1.15.1...v1.15.2
 [1.15.1]: https://github.com/lumeland/lume/compare/v1.15.0...v1.15.1
 [1.15.0]: https://github.com/lumeland/lume/compare/v1.14.2...v1.15.0
 [1.14.2]: https://github.com/lumeland/lume/compare/v1.14.1...v1.14.2